pyproject.toml
[tool.poetry]
name = "zeebot"
version = "0.1.0"
description = ""
authors = ["zhufree <zhufree2013@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.5"
nonebot-adapter-onebot = "^2.1.5"
nonebot-plugin-apscheduler = "^0.2.0"
nonebot-plugin-epicfree = "^0.2.0"
nonebot-plugin-simplemusic = "^0.1.4"
nonebot-plugin-easyCommand = "^1.1.2"
nonebot-bison = "^0.6.1"
nb-cli = "^0.6.7"
pyquery = "^1.4.3"


[tool.nonebot]
plugins = ["nonebot_plugin_simplemusic", "nonebot_plugin_epicfree", "nonebot_bison", "nonebot_plugin_easyCommand", 
"nonebot_plugin_wordcloud", "nonebot_plugin_everyday_en"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
